'compile' now handles the renaming of module's bad global functions.
After builing vlc's modules, 'compile' goes over each of the modules
and rename the following functions if they are present:
    -AccessOpen
    -AccessClose
    -StreamOpen
    -StreamClose
    -DemuxOpen
    -DemuxClose
    -OpenFilter
    -CloseFilter
    -Open
    -Close
    -vlc_entry__*
    -vlc_entry_copyright__*
    -vlc_entry_licence__*

It renames those is by adding the correct module name at the end of the functions.

   This on vlc_android was done by using "objcopy --redefine-syms" but couldn't be used
here as nacl's toolchain doesn't ship a 'objcopy' implementation for 'pnacl bitcode' (not ELF).

     The workaround found here is to use a combination of "pnacl-opt"
that translate 'pnacl bitcode' to llvm intermediate representation (IR) back and forth
and sed to replace the symbols as need be.

    The way we match a function to be replaced is if it's preceded by either a '@' or a '.'.
In certain, produced, 'llvm IR' files we can find references to functions we need to replace inside a string without being preceded by a '@' nor a '.'. The impact of those cases is yet unkown and cannot be resolved using 'sed' as the length of said string is written somewhere else before on the line.

   On the current state of testing this issue didn't trigger any change in bahaviour.

   It is likely that thoses strings are there for debuging purpuse and won't impact the compilation.