Update/Delete/Split MVK the patch.
The MKV patch contains changes present in a part of a VLC commit
on the original VLC repository pushed after our fork point.
This part of the patch have been therefore removed from
the patch and the original patch have been copied
in the patch folder using the name catchup-vlc-0001.patch.

The rest of our MKV patch have been replaced by and other one written by
Filip Roséen. It is not yet the final version of said patch which is the
reason why it has not been merged on VLC's repository.