Create a separate shell script for renaming symbols in modules.
The part of the 'compile' script that handled
the renaming of symbols in VLC's modules have been moved in it's
own shell script.

This allows us to invoke it in a Makefile so that we can take
advantage of the fact the Makefiles handle multithreading natively.

Therefore the Makefile has now a rule for '%.a.correct' files being
dependant on module file '.a' and this rule calls our script.

Where the script used to be, we ask of Makefile to build all the
'%.a.corrected' files and use -j`nproc` to make it faster.

The reason why we do all of this is that in the previous
implementation if one of the module were to fail it wouldn't have
stop the compilation and the user might have missed it considering
the sheer number of debugging information.

This is corrected.